* AddressManager/Controller.m         * AddressManager/DragDropMatrix.h         * AddressManager/DragDropMatrix.m         * Frameworks/AddressView/ADPersonView.m         * Frameworks/AddressView/ADSinglePropertyView.m         * Frameworks/Addresses/ADPerson.m         * Goodies/adserver/adserver.m           (unsigned) int -> NS(U)Integer and float -> CGFloat           and get rid of some other warnings emitted by clang
git-svn-id: svn+ssh://svn.savannah.nongnu.org/gap/trunk@2221 d25b4930-3292-e311-b074-0010c6de5da8

+++ system-apps/Addresses/AddressManager/Controller.m
@@ -911,8 +911,8 @@
  *  Browser delegate methods
  */
 
-- (int) browser: (NSBrowser*) sender
-numberOfRowsInColumn: (int) column
+- (NSInteger) browser: (NSBrowser*) sender
+numberOfRowsInColumn: (NSInteger) column
 {
   NSArray *groupnames = [self groupNames];
 
@@ -986,8 +986,8 @@ numberOfRowsInColumn: (int) column
 
 - (void) browser: (NSBrowser*) sender
  willDisplayCell: (id) cell
-	   atRow: (int) row
-	  column: (int) column
+	   atRow: (NSInteger) row
+	  column: (NSInteger) column
 {
   [cell setFont: [NSFont systemFontOfSize: [NSFont systemFontSize]]];
   if(column == 0)
